refactor: #Y-57 Унификация названий файлов
