--- packages/client/src/components/Container/index.tsx
@@ -1,3 +1,0 @@
-import { Container } from './Container'
-
-export default Container
