--- packages/client/src/components/FormInput/index.ts
@@ -1,3 +1,0 @@
-import { FormInput } from './FormInput'
-
-export default FormInput
