+++ packages/client/src/components/container/index.tsx
@@ -1,0 +1,3 @@
+import { Container } from './container'
+
+export default Container
