+++ packages/client/src/components/form/index.tsx
@@ -1,3 +1,3 @@
-import { Form } from './Form'
+import { Form } from './form'
 
 export default Form
