+++ packages/client/src/components/formInput/index.ts
@@ -1,0 +1,3 @@
+import { FormInput } from './formInput'
+
+export default FormInput
