+++ packages/client/src/components/menu/index.ts
@@ -1,1 +1,1 @@
-export { Menu } from './Menu'
+export { Menu } from './menu'
