+++ packages/client/src/components/startModal/gameRule/gameRule.tsx
