+++ packages/client/src/lib/chess/core/index.ts
@@ -1,4 +1,4 @@
-export * from './Board'
-export * from './Game'
-export * from './Player'
-export * from './Tank'
+export * from './board'
+export * from './game'
+export * from './player'
+export * from './tank'
