--- packages/client/src/pages/AuthPage/SignUp/index.ts
@@ -1,3 +1,0 @@
-import { SignUp } from './SignUp'
-
-export default SignUp
