--- packages/client/src/pages/AuthPage/index.ts
@@ -1,3 +1,0 @@
-import { AuthPage } from './AuthPage'
-
-export default AuthPage
