+++ packages/client/src/pages/authPage/index.ts
@@ -1,0 +1,3 @@
+import { AuthPage } from './authPage'
+
+export default AuthPage
