+++ packages/client/src/pages/authPage/signIn/index.ts
@@ -1,0 +1,3 @@
+import { SignIn } from './signIn'
+
+export default SignIn
