+++ packages/client/src/pages/authPage/signUp/index.ts
@@ -1,0 +1,3 @@
+import { SignUp } from './signUp'
+
+export default SignUp
