+++ packages/client/src/pages/authPage/signUp/signUp.module.scss
