--- packages/client/src/pages/gamePage/GamePage.module.scss
@@ -1,8 +1,0 @@
-.game{
-    display: flex;
-    gap: 16px;
-}
-
-.board{
-    margin: 20px;
-}
