--- packages/client/src/pages/gamePage/components/Info/index.ts
@@ -1,1 +1,0 @@
-export * from './Info'
