--- packages/client/src/pages/gamePage/components/TankController/index.ts
@@ -1,1 +1,0 @@
-export * from './TankController'
