--- packages/client/src/pages/gamePage/components/TankInfo/index.ts
@@ -1,1 +1,0 @@
-export * from './TankInfo'
