+++ packages/client/src/pages/gamePage/components/info/index.ts
@@ -1,0 +1,1 @@
+export * from './info'
