+++ packages/client/src/pages/gamePage/components/info/info.module.scss
