+++ packages/client/src/pages/gamePage/components/tankController/index.ts
@@ -1,0 +1,1 @@
+export * from './tankController'
