+++ packages/client/src/pages/gamePage/components/tankController/tankController.module.scss
