+++ packages/client/src/pages/gamePage/components/tankInfo/index.ts
@@ -1,0 +1,1 @@
+export * from './tankInfo'
