+++ packages/client/src/pages/gamePage/components/tankInfo/tankInfo.module.scss
