+++ packages/client/src/pages/gamePage/index.ts
@@ -1,3 +1,3 @@
-import { GamePage } from './GamePage'
+import { GamePage } from './gamePage'
 
 export default GamePage
