+++ packages/client/src/pages/mainPage/mainPage.tsx
