+++ packages/client/src/pages/profilePage/components/avatar/avatar.tsx
