+++ packages/client/src/pages/profilePage/components/avatar/index.tsx
@@ -1,3 +1,3 @@
-import { Avatar } from './Avatar'
+import { Avatar } from './avatar'
 
 export default Avatar
