+++ packages/client/src/pages/profilePage/components/formPassword/index.tsx
@@ -1,3 +1,3 @@
-import { FormPassword } from './FormPassword'
+import { FormPassword } from './formPassword'
 
 export default FormPassword
